全流程

+++ Assets/Scripts/Buff/BaseBuff/Buff/Buff_1007.cs
@@ -1,15 +1,15 @@
 public class Buff_1007 : BuffBase
 {
-	public override void OnStart(Entity entity,int Value)
+	public override void OnStart(Entity entity,float Value)
 	{
 	}
-	public override void OnEnd(Entity entity,int Value)
+	public override void OnEnd(Entity entity,float Value)
 	{
 	}
-	public override void OnAdd(Entity entity,int Value)
+	public override void OnAdd(Entity entity,float Value)
 	{
 	}
-	public override void OnRemove(Entity entity,int Value)
+	public override void OnRemove(Entity entity,float Value)
 	{
 	}
 }
